Implemented central config file for all services
The shared module now includes a general load_ctf_config function that
returns a dict including all configuration values that is supposed to be
available for all services. Currently it's using a YAML config (defaults
to config.yaml in the current folder) that can be supplied via the run_*
service initialization functions.

The implementation of this function and the file format can be changed
in the future, as long as the paradigms (dict with coresponding key
names) don't change.